create gatsby-plugin-git-info
This local plugin is a fork I created of gatsby-transformer-gitinfo,
which was originally used but suffered from serious performance issues.
On every call to onCreateNode, the library would completely reindex the
entire git repo in order to retrieve the latest commit for a single file.

This new plugin instead performs everything during the onPreInit stage.
It first matches all markdown files (as these are currently the only
ones which need to display a 'Last updated' tooltip) and then caches the
latest commit for each file. Then, at build time it simply retrieves this
info with a lookup.

On my machine I saw a significant speedup on the 'source and
transform nodes' stage, going from taking +120s to now only taking
~40s!